requirements.txt
astroid==2.4.2
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dropbox==10.2.0
et-xmlfile==1.0.1
flake8==3.8.3
Flask==1.1.2
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openpyxl==3.0.4
Pillow==7.2.0
pkg-resources==0.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.3
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
